🙉🎢 Checkpoint ./commands/moderation/warn.js:29885963/3204 ./commands/moderation/warnings.js:29885963/1411 ./commands/moderation/clear.js:29885963/4148

+++ commands/moderation/clear.js
@@ -1,53 +1,32 @@
-const Discord = require("discord.js");
-const { MessageEmbed } = require("discord.js");
-const { Color } = require("../../config.js");
-
 module.exports = {
   name: "clear",
-  aliases: ["purge", "clearmsgs"],
-  description: "Clear Your Messages!",
-  usage: "Clear <Message Amount>",
+  aliases:["c", "purge"],
+  category: "moderation",
+  description: "Delete bulk messages with 1 command",
   run: async (client, message, args) => {
-    //Start
-    message.delete();
-    if (!message.member.hasPermission("MANAGE_MESSAGES"))
-      return message.channel.send(
-        "You Don't Have Permission To Use This Command!"
-      );
+    
+            
+    if (message.deletable) {
+        message.delete();
+    }
 
-    if (!args[0])
-      return message.channel.send(`Please Give Me Amounts Of Messages!`);
+    if (!message.member.hasPermission("MANAGE_MESSAGES")) {
+        return message.reply("Missing Permissions!").then(m => m.delete(5000));
+    }
 
-    if (isNaN(args[0]))
-      return message.channel.send(`Please Give Me Number Value!`);
+    if (isNaN(args[0]) || parseInt(args[0]) <= 0) {
+        return message.reply("This is not a number").then(m => m.delete(5000));
+    }
 
-    if (args[0] < 4)
-      return message.channel.send(
-        `You Can Delete ${args[0]} By Your Self Its Not Too Many Messages!`
-      );
+    let deleteAmount;
+    if (parseInt(args[0]) > 100) {
+        deleteAmount = 100;
+    } else {
+        deleteAmount = parseInt(args[0]);
+    }
 
-    if (args[0] > 100)
-      return message.channel.send(
-        `I Can't Delete ${args[0]} Because Of Discord Limit!`
-      );
-
-    let Reason = args.slice(1).join(" ") || "No Reason Provided!";
-
-    message.channel.bulkDelete(args[0]).then(Message => {
-      let embed = new Discord.MessageEmbed()
-        .setColor(Color)
-        .setTitle(`Messages Deleted!`)
-        .addField(`Moderator`, `${message.author.tag} (${message.author.id}`)
-        .addField(`Channel`, `${message.channel.name} (${message.channel.id}`)
-        .addField(`Deleted Messages`, `${Message.size}`)
-        .addField(`Reason`, `${Reason}`)
-        .setFooter(`Requested by ${message.author.username}`)
-        .setTimestamp();
-      return message.channel
-        .send(embed)
-        .then(msg => msg.delete({ timeout: 0 }));
-    });
-
-    //End
+    message.channel.bulkDelete(deleteAmount, true)
+    .catch(err => message.reply(`Something went wrong... ${err}`));
+    
   }
-};
+}
